Remove MODERATION-GUIDELINES
The moderation guidelines as they have been established and how
they are enforced are in violation in contributors rights as set
by all the major jurisdictions from where are historical or active
individual contributors or organizational stakeholders to the
bitcoin core project.

For now, remove the MODERATION-GUIDELINES.md until there is legal
clarity on the role of contributor consent w.r.t any hypothetical
moderation rules.

It should be noted as a matter of fact that the MODERATION-GUIDELINES
were adopted afterthe CoreDev.tech of April 8-11 in Berlin, Germany,
an invitation-only in-person meetings, where not all the historical
or active contributors to bitcoin consensus rules were present.